global debug

+++ src/main/java/ru/skypro/homework/dto/comment/Comment.java
@@ -10,6 +10,6 @@ public class Comment {
     private int author;
     private String authorImage;
     private String authorFirstName;
-    private int createdAt;
+    private long createdAt;
     private String text;
 }
